+++ platforms/ide/tooling-api/src/main/java/org/gradle/tooling/events/problems/internal/DefaultOffsetInFileLocation.java
@@ -19,22 +19,22 @@ package org.gradle.tooling.events.problems.internal;
 import org.gradle.tooling.events.problems.OffsetInFileLocation;
 
 public class DefaultOffsetInFileLocation extends DefaultFileLocation implements OffsetInFileLocation {
-    private final int offset;
-    private final int length;
+    private final long offset;
+    private final long length;
 
-    public DefaultOffsetInFileLocation(String path, int offset, int length) {
+    public DefaultOffsetInFileLocation(String path, long offset, long length) {
         super(path);
         this.offset = offset;
         this.length = length;
     }
 
     @Override
-    public int getOffset() {
+    public long getOffset() {
         return offset;
     }
 
     @Override
-    public int getLength() {
+    public long getLength() {
         return length;
     }
 }
